hydroreport: make underlying data visible
We want it so that we can produce other info on the reports page.
Also, add MeterSampleDir.OpenRange factored out from the hydroreport
code.
